requirements.txt
plexapi
tinydb
pyyaml
bjoern
bottle